Finished hw01, started hw02
I added a PDF of hw01 to the git as well so people can easily read it.
I have not went through and edited it. The plan is to get as much done
as possible and then go back through with a fine tooth comb. I made
hw01 a template so I can easily start another homework document, which
is what I did for hw02.